rename and move updateShareMemberPermissions
